shill: WiFi: Enable WiFi debugging for persistent failures
If a service has recent connection failures, enable wpa_supplicant
debugging for the duration of any connection attempt to this
service.

BUG=chromium-os:36469
TEST=Unit tests + Manual: repeat failing connection attempt to
WPA-PSK AP with the wrong passphrase.  Use command-line "wpa_debug"
command to note that during successive connection attempts
wpa_supplicant is set to the "debug" debugging level, and when the
system returns to an idle state it returns to the "info" debugging
level.

Change-Id: Ibef464fe0599f511a937c09a727c703ebdf1ebbf
Reviewed-on: https://gerrit.chromium.org/gerrit/39882
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>