Powerwash: Collect Stats
When the device is powerwashed, keep track of how many times it's been
powerwashed sequentially (i.e. in between complete recoveries of the
device and inbetween switching to/from dev mode.).

After every powerwash send to metrics the number of times it's been done
on this device.

BUG=chromium-os:37110
TEST=Tested on stumpy and verified the stats end up in about:histograms
     and the count increases on each powerwash.

Change-Id: I92344fb7e70ac2a6190a0954b6298849d4608897
Reviewed-on: https://gerrit.chromium.org/gerrit/39971
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>