Convert the send_boot_metrics script to an upstart job.
The send_boot_metrics script was invoked in the background from the
boot-complete job.  There was no reason for the script's privileged
position other than the momentum of history, so the script is moving
to be a regular upstart job depending on 'system-services' instead.

BUG=None
TEST=Boot system, check chrome://histograms for relevant metrics

Change-Id: Iee3f11fb8b6e24dfcd76aab884c505e59be39c99
Reviewed-on: https://gerrit.chromium.org/gerrit/40025
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>