power: Move input code from powerm to powerd.
This moves powerm's Input class to powerd and adds a new
InputController class that's responsible for handling input
events.  It deletes the powerd-liveness code from powerm.
At this point, powerm is only responsible for suspending the
system and shutting it down; after that code is moved to
powerd, powerm can be removed.

BUG=chromium-os:36804
TEST=manual:
     - power button works as before
     - system shuts down on lid-close on login screen
     - system suspends on lid-close after logging in
     - reopening lid just after closing cancels suspend
     - idle-suspend works as before
     - hitting a key just after idle-suspend cancels suspend

Change-Id: I91aba04fe1ade45c1d7ae2e0e9fe70a5828f4e75
Reviewed-on: https://gerrit.chromium.org/gerrit/40307
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>