power: Fix bug where user activity doesn't abort suspend.
It looks like there's a typo in Daemon::SetIdleState().  In
the going-to-active transition, the method sets the state to
BACKLIGHT_ACTIVE and then tests whether the state is
BACKLIGHT_SUSPENDED before canceling an in-progress suspend
attempt.  Presumably it should be testing whether the
previous state was suspended instead.

This change also removes an unused Daemon::power_state_
member.

BUG=none
TEST=manual: ran set_short_powerd_timeouts.  tailed
     /var/log/power_manager/powerd.LATEST and hit a key just
     after a suspend request was sent.  checked that the
     screen stayed on and the system didn't suspend.
     without this change, powerd turns the screen on in
     response to the keypress but powerm suspends anyway.

Change-Id: I4669215867659949299e6d7d630f36060cb6e4c9
Reviewed-on: https://gerrit.chromium.org/gerrit/40287
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>