Add zram field trial (/etc/init).
This adds the zram (compressed swap) field trial (experiment).
It consists of three CLs: this one, one CL in crosh (to opt-out
of the experiment when manually enabling zram), and one CL in
Chrome to set up the UMA stuff.

The experiment groups are selected by hashing the MAC address
of the wifi adapter on the device.  Ideally we would assign
users to groups, because we believe that the presence of compressed
swap may change the usage pattern.  Unfortunately it is difficult
to turn swap off without rebooting, so we settle for device-based
groups as an approximation.

More details in the code.

BUG=chromium-os:37583
TEST=manual testing for a number of combinations (not all)

Change-Id: I457601a3b96464c2bc4e2fdbd305ca2e9bec1fbb
Reviewed-on: https://gerrit.chromium.org/gerrit/40354
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>