powerd: Add support for resume backlight level.
To speed up resume, we plan to turn on the panel before idle suspend.
In turn, this requires that we turn off the backlightbefore turning
on the display before idle suspend. To achieve that, we use the new
resume brightness value in the kernel from powerd.

This requires the following patch which adds resume brightness
support to the kernel:
https://gerrit.chromium.org/gerrit/#/c/37974/

BUG=chrome-os-partner:13364
TEST=By hand, on link: idle resume is now fast.
TEST=Normal resume is still fast.

Change-Id: I3e3f26e940b9adb51f5c6a2d852bc076df10729e
Reviewed-on: https://gerrit.chromium.org/gerrit/40971
Reviewed-by: Simon Que <sque@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>