Add zram field trial (/etc/init) with correct file owners
This adds the zram (compressed swap) field trial (experiment).
It consists of three CLs: this one, one CL in crosh (to opt-out
of the experiment when manually enabling zram), and one CL in
Chrome to set up the UMA stuff.

The experiment groups are selected by hashing the MAC address
of the wifi adapter on the device.  Ideally we would assign
users to groups, because we believe that the presence of compressed
swap may change the usage pattern.  Unfortunately it is difficult
to turn swap off without rebooting, so we settle for device-based
groups as an approximation.

This replaces #40354, which had to be reverted because it
created files owned by root in /home/chronos.

BUG=chromium-os:37583
TEST=manual testing for a number of combinations (not all)

Change-Id: I618a4726b6b8f2e7dfe681a7516b480d1f10f177
Reviewed-on: https://gerrit.chromium.org/gerrit/40999
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>