power: Fix spurious low battery shutdown on energy-reporting batteries.
When a battery is initially plugged or the system comes up after boot /
suspend, there is some chance for sysfs battery metrics to report 0 for
a while. This is handled correctly for charge-reporting batteries, but
for energy-reporting batteries, we never assign battery_percentage
because we return early from GetPowerStatus. This can lead to incorrect
shutdown due to battery_percentage being zero.

Fix is to set battery_percentage to indicate an invalid / uncalculated
value in the case where voltage is not available, as we do on
charge-reporting batteries when needed metrics are not available.

BUG=chrome-os-partner:17166.
TEST=manual. Plug battery + pull AC many times, verify no incorrect
shutdown due to low battery.

Change-Id: I793bf28fff7af5ab090c31270cd353fdf837d74d
Reviewed-on: https://gerrit.chromium.org/gerrit/41211
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>