shill: Include pre-enabled technologies in EnabledTechnologies property.
This CL changes the EnabledTechnologies property of the Manager
interface to include the technologies of 'pre-enabled' devices that have
been detected by shill but may not have finished initialization and
created a Device object. This allows the CrOS network UI to show the
the technologies of pre-enabled devices as 'initializing'.

BUG=chromium-os:37877
TEST=Tested the following:
1. Build and run unit tests.
2. Verify that CrOS network UI shows the cellular technology as
   initializing shortly after the modem is detected but before it
   is fully initialized.

Change-Id: I6a98527750eda06cb533908a22c0cca09cea5cad
Reviewed-on: https://gerrit.chromium.org/gerrit/41307
Commit-Queue: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>