lockbox-cache.conf: add a task to emit a lockbox cache
In order to make the lockbox data accessible as early as possible,
we can verify using the NVRAM data collected by mount-encrypted staged
in /tmp/lockbox.nvram.  This task uses that file and generates a
file in /var/run/lockbox which contains a cache of the install_attributes.pb
iff the file validated against the boot-time nvram data.

In the cases of VMs or non-chrome firmware and there is no NVRAM file,
we just pretend that the lockbox is validated.

BUG=chromium-os:37367
TEST=emerges and runs lockbox-cache as soon as possible generating the
cache when it is legit, but adding a byte to the lockbox content shows
it doesn't regenerate.

Change-Id: I8616095b54b44ea392630dad00769d5fcc383119
Reviewed-on: https://gerrit.chromium.org/gerrit/41538
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>