shill: Manager: Implement FindMatchingService
Find an in-memory Service given a dictionary of attributes.  The
Manager will return the first Service that matches these attributes.
While here, fix an old bug where Service::Configure enumerated the
input uint32 arguments with the intention of configuring the
"Priority" property, except that "Priority" is an int32.

BUG=chromium-os:34525
TEST=Unit tests, manual using python script to call GetMatchingService
with a known service name

Change-Id: Ib093b257c4217796ae5aa13bf090c1234df70c7d
Reviewed-on: https://gerrit.chromium.org/gerrit/41554
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>