Crosh: strip trailing whitespace from command name
I'm not entirely sure why, but sometimes the space between a command and
its parameters is included into the command's name causing Crosh to
reject that command.  Only some machines that I've tested are effected,
but just making sure there are no spaces in the command string seems to
fix it in any case.

For example:
    crosh> tracepath cmooney.mtv
    Unknown command 'tracepath '

BUG=chromium-os:38028
TEST=manually ran this copy of Crosh on an effected computer and was
able to repeatedly run the troublesome commands.  Commands without
parameters such as "shell" still worked too.

Change-Id: Icfaa939671f04a63f3961e87b203f335de5246e1
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41641
Reviewed-by: Robert Ginda <rginda@chromium.org>