Let the EC handle saving and restoring KBL state on suspend/resume
Currently the KBC is turning down its target percents on suspend to dim the
backlight. This gives a nice transition, but since is doesn't save/restore these
values there are some anomolous behaviours on resume since the EC also does
save/restore. When implementing save/restore in s/w the transitions were not
preserved, since the device doesn't delay suspending for the backlight. Thius
led to situations where the transition was partly done, device suspended, so the
EC restored a weird brightness level. The fix for this was moving to an instant
transition for suspend.

If we are going to use an instant transition anyway, there is no need to fiddle
with the target percents in s/w and just let the EC handle the save/restore for
suspend/resume. This CL short ciruits the adjustment logic when going to/from
the suspend state and lets the EC do its thing.

BUG=chrome-os-partner:17362
TEST=Took the device into an environment where the ALS turned on the light,
     suspended the device and confirmed on resume the light returned to the
     expected level.
     Used the user controls to set the backlight level, suspended the device and
     confirmed on resume the light returned to the expected level.

Change-Id: Ic8d89c5e33eff5b43ad01a8862f45beb25517695
Reviewed-on: https://gerrit.chromium.org/gerrit/41839
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>