Make swap-exp a task to eliminate races, + extra fix
When swap-exp.conf is not a task, it can execute in parallel with
swap.conf, leading to races.

In addition, remove the code that makes swap-exp run only when
NOT in developer mode.  I have observed hang situations that may
be attributed to crossystem, but they are hard to reproduce.
The UMA histogram site already filters out devices in dev mode,
so it is not necessary to do it here.

BUG=chromium-os:38112
TEST=manually tested
BRANCH=none

Change-Id: If4cb843f2e3889759608a2d65f44c4194d6d32d9
Reviewed-on: https://gerrit.chromium.org/gerrit/41779
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>