shill: cellular: Update Scanning property in CellularCapabilityUniversal.
This CL updates the Device.Scanning property in
CellularCapabilityUniversal such that the property is set to true if
there is a ongoing network scan or the modem is enabled but not yet
registered. This allows the network UI to show the initializing/scanning
animation before the modem is registered to a network.

BUG=chrome-os-partner:17389
TEST=Tested the following:
1. Build and run unit tests.
2. Observe that the network UI starts showing the animation as soon as
   the modem is being enabled and stops the animation after the modem
   finishes registration with a network (i.e. a cellular service is
   present).

Change-Id: I75532e81451eff38dc13fd7c37a5be203ce3f3f5
Reviewed-on: https://gerrit.chromium.org/gerrit/41837
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>