powerd_suspend: write time in foreground
When we wake up from suspend, if we're doing a test, we write the current time
to a file. This write was done in the background, but we need to do it in the
foreground to make sure we block on the creation of the file.

BUG=chrome-os-partner:17226
TEST=suspend and resume lumpy. run power_Resume

Change-Id: Icf2b4b0e4def75988bf34b04095af6c3d6f7d994
Reviewed-on: https://gerrit.chromium.org/gerrit/41962
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>