Start swap-exp.conf earlier so Chrome can start trial
Before this CL, swap-exp.conf runs after the Chrome code that
starts the field trial, so the experiment misses the first boot
session (which could be quite long).

The execution time of the script is small, and it is negligible
after the initial setup, so impact on boot time is acceptable.

BUG=chromium-os:37583
TEST=manual
BRANCH=none

Change-Id: I8f2fb3625389995f72eae01a1c31a78243b47145
Reviewed-on: https://gerrit.chromium.org/gerrit/42029
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>