power: Fix timeout double-free in AsyncFileReader.
AsyncFileReader::UpdateState() contains a path where it
returns false (to cancel the timeout that ran it) but
doesn't reset the variable containing the timeout's ID to 0.
If this path is taken and the AsyncFileReader is destroyed
before a new timeout has been registered, it would result in
GLib trying to remove the old, already-canceled timeout.

I'm speculatively correcting this since I saw one of
AmbientLightSensor's tests fail with a "corrupted
double-linked list" message on a bot and I've already fixed
all of the other timeout-related bugs that I've been able to
find in this code.

BUG=none
TEST=tests still pass locally
TBR=sque@chromium.org

Change-Id: I488e5ebf78304d9faaa770d4d0104638bf7680cc
Reviewed-on: https://gerrit.chromium.org/gerrit/42392
Commit-Queue: Daniel Erat <derat@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>